Move the nodeset generation out of the app
* remove need for bluepy
* add tests; output is the same as when using bluepy
